last commit?

+++ src/executor/child_routine.c
@@ -3,7 +3,7 @@
 /*                                                        :::      ::::::::   */
 /*   child_routine.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: martorre <martorre@student.42.fr>          +#+  +:+       +#+        */
+/*   By: rbarbier <rbarbier@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/13 13:04:40 by rbarbier          #+#    #+#             */
 /*   Updated: 2024/03/13 14:13:31 by martorre         ###   ########.fr       */
@@ -63,16 +63,20 @@ void	is_global_cmd(t_cmd *cmd, t_pipe data, char **envp)
 	msg_exit(cmd->args[0], 0, ERR_CMD_NOT_FOUND, 127);
 }
 
-void	unlink_heredoc(t_rd *redir)
+void	unlink_heredoc(t_cmd *cmd)
 {
 	t_rd	*tmp;
 
-	tmp = redir;
-	while (tmp)
+	tmp = cmd->redir;
+	while (cmd)
 	{
-		if (tmp->type == HEREDOC)
-			unlink(tmp->file);
-		tmp = tmp->next;
+		while (tmp)
+		{
+			if (tmp->type && tmp->type == HEREDOC)
+				unlink(tmp->file);
+			tmp = tmp->next;
+		}
+		cmd = cmd->next;
 	}
 }
 
@@ -80,9 +84,10 @@ void	child(t_pipe data, t_cmd *cmd, t_env **env, t_env **exp)
 {
 	char	**envp;
 
-	if (/*cmd->args == NULL*/cmd->args == 0 || cmd->args[0] == 0 || cmd->args[0][0] == 0)
-		return (ft_fprintf(2, "minishell: %s: command not found\n", cmd->args[0]), exit(127));
-		//return (ft_fprintf(2, NOTFOUND), exit(127));
+	if (cmd->args == NULL)
+		return ;
+	if (cmd->args[0][0] == 0)
+		return (ft_fprintf(2, NOTFOUND, cmd->args[0]), exit(127));
 	envp = env_to_array(env);
 	init_signals(0);
 	get_files_redir(cmd->redir, &data);
@@ -92,5 +97,4 @@ void	child(t_pipe data, t_cmd *cmd, t_env **env, t_env **exp)
 		exit(exit_status(0));
 	is_local_cmd(cmd, data, envp);
 	is_global_cmd(cmd, data, envp);
-	unlink_heredoc(cmd->redir);
 }
